Bug fixes
Issues with the validation (again) and changed the div's so that it looks better when on a small device.

+++ 04-a-big-barbecue/api/Entities/BarbecueDetails.cs
@@ -12,7 +12,7 @@ namespace BarbecueCost.Api.Entities
         public int Guests { get; set; }
 
         [JsonProperty("budget")]
-        public int Budget { get; set; }
+        public double Budget { get; set; }
 
         [JsonProperty("barbecueItems")]
         public List<BarbecueItem> BarbecueItems { get; set; }
